requirements.txt
numpy
Flask
onnx
Pillow